allow nesting of WITH-FACETs for the same facet
This used to signal an error:

    (with-facet (a (m 'array :direction :input))
      (with-facet (a (m 'array :direction :io))
        ...))

Not anymore. This allow unsuspecting, general purpose functions (think
AXPY!) to correctly deal with cases where some of the arguments are
the same MAT object (as in (AXPY! 1 X X)).

Nesting WITH-FACETs for different facets is still an error and so is
having a read/write conflict between threads.