add CUDA-HOST-ARRAY facet
It is like the FOREIGN-ARRAY facet but enables quick copying to/from
the GPU by registering the memory with cuMemHostRegister.
Unregistering happens when the CUDA-HOST-ARRAY is destroyed which is
done automatically by WITH-CUDA*.

With *FOREIGN-ARRAY-STRATEGY* :CUDA-HOST, MATs will be allocated as
static vectors and any copying to/from the GPU will go via the
CUDA-HOST-ARRAY facet (even if it has to be created). To enable this
behavior:

    (let ((*foreign-array-strategy* :cuda-host))
      (with-cuda* ()
        ...))

Also, these *FOREIGN-ARRAY-STRATEGY* values were renamed:

- :PIN-BACKING-ARRAY -> :PINNED
- :STATIC-BACKING-ARRAY-P -> :STATIC

:DYNAMIC is still the same and :CUDA-HOST was added.

Also, SELECT-COPY-SOURCE-FOR-FACET* is now allowed to return
non-up-to-date facets in which case the returned facet will be updated
first.