factor out VEC
A VEC is to a MAT what an underlying simple vector (also known as
backing array) is to a lisp array: a non-displaced, simple, 1d
storage. It has three independent facets: LISP-VECTOR, STATIC-VECTOR
and CUDA-VECTOR. A MAT object has an internal VEC and MAT's facets
alias facets of VEC. The same VEC may belong to several MATs.